Move setup package
