+++ cmd/saucectl/saucectl.go
@@ -6,7 +6,7 @@ import (
 	initCmd "github.com/saucelabs/saucectl/cli/command/init"
 	"github.com/saucelabs/saucectl/cli/command/run"
 	"github.com/saucelabs/saucectl/cli/command/signup"
-	"github.com/saucelabs/saucectl/cli/setup"
+	"github.com/saucelabs/saucectl/internal/setup"
 	"github.com/saucelabs/saucectl/internal/version"
 	"os"
 	"time"
